requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-telegram-bot==21.6
python-dotenv==1.0.1
